backend/app.py
import os
# ---- OPTIONAL: Silence TensorFlow oneDNN logs ----
os.environ["TF_ENABLE_ONEDNN_OPTS"] = "0"

from flask import Flask, request, jsonify
from tensorflow.keras.models import load_model
import numpy as np
from PIL import Image
from io import BytesIO
from flask_cors import CORS
import logging
import traceback
from datetime import datetime
import librosa
import soundfile as sf
from flask_sqlalchemy import SQLAlchemy

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

app = Flask(__name__)
CORS(app)

# Configuration
ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif', 'bmp', 'tiff'}
MAX_FILE_SIZE = 10 * 1024 * 1024  # 10MB

BASE_DIR = os.path.dirname(os.path.abspath(__file__))
MODEL_PATH = os.path.join(BASE_DIR, "parkinson_mobilenet_augmented.keras")
CLASS_NAMES = ["Healthy", "Parkinson"]

# =========================
# Database Configuration (FIXED)
# =========================
DB_USER = "root"
DB_PASSWORD = "root"
DB_HOST = "localhost"
DB_NAME = "parkinson_db"   # ✅ REAL database name

app.config['SQLALCHEMY_DATABASE_URI'] = (
    f"mysql+mysqlconnector://{DB_USER}:{DB_PASSWORD}@{DB_HOST}/{DB_NAME}"
)
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False

db = SQLAlchemy(app)

# --- Models ---
# --- Models ---
class Doctor(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    username = db.Column(db.String(80), unique=True, nullable=False)
    password = db.Column(db.String(120), nullable=False)
    name = db.Column(db.String(100), nullable=False)
    specialization = db.Column(db.String(100), default="Neurologist")

class Patient(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    username = db.Column(db.String(80), unique=True, nullable=False)
    password = db.Column(db.String(120), nullable=False)
    name = db.Column(db.String(100), nullable=False)
    age = db.Column(db.Integer, nullable=True)
    contact = db.Column(db.String(20), nullable=True)

class Prediction(db.Model):
    id = db.Column(db.Integer, primary_key=True)
    patient_id = db.Column(db.Integer, db.ForeignKey('patient.id'), nullable=False)
    type = db.Column(db.String(20), nullable=False)
    result = db.Column(db.String(50), nullable=False)
    confidence = db.Column(db.Float, nullable=False)
    timestamp = db.Column(db.DateTime, default=datetime.utcnow)

    # Relationship
    patient = db.relationship('Patient', backref=db.backref('predictions', lazy=True))

# Initialize DB
with app.app_context():
    try:
        db.create_all()
        
        # Seed Doctor
        if not Doctor.query.filter_by(username='doctor').first():
            db.session.add(Doctor(username='doctor', password='password', name="Dr. Strange"))
            db.session.commit()
            print("Doctor seeded.")

        # Seed Patient
        if not Patient.query.filter_by(username='patient').first():
            db.session.add(Patient(username='patient', password='password', name="John Doe", age=65, contact="1234567890"))
            db.session.commit()
            print("Patient seeded.")
            
        logger.info("Database initialized successfully")
    except Exception as e:
        logger.error(f"Database initialization failed: {e}")

# Load model
try:
    if os.path.exists(MODEL_PATH):
        model = load_model(MODEL_PATH)
        logger.info("Model loaded successfully")
    else:
        logger.error(f"Model file not found at {MODEL_PATH}")
        model = None
except Exception as e:
    logger.error(f"Error loading model: {e}")
    model = None

def allowed_file(filename):
    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS

def validate_image(file):
    try:
        file.seek(0, 2)
        file_size = file.tell()
        file.seek(0)

        if file_size > MAX_FILE_SIZE:
            return False, "File too large"

        img = Image.open(BytesIO(file.read()))
        file.seek(0)

        w, h = img.size
        if w < 50 or h < 50:
            return False, "Image too small"
        if w > 5000 or h > 5000:
            return False, "Image too large"

        return True, "Valid image"
    except Exception as e:
        return False, str(e)

@app.route("/health", methods=["GET"])
def health_check():
    return jsonify({
        "status": "healthy" if model else "unhealthy",
        "model_loaded": model is not None,
        "timestamp": datetime.utcnow().isoformat()
    })

@app.route("/predict-test", methods=["GET"])
def test():
    return jsonify({"message": "Backend is working"})

@app.route("/predict", methods=["POST"])
def predict():
    if model is None:
        return jsonify({"error": "Model not loaded"}), 503

    if 'file' not in request.files:
        return jsonify({"error": "No file uploaded"}), 400

    file = request.files['file']

    if not allowed_file(file.filename):
        return jsonify({"error": "Invalid file type"}), 400

    valid, msg = validate_image(file)
    if not valid:
        return jsonify({"error": msg}), 400

    img = Image.open(BytesIO(file.read())).convert("RGB")
    img = img.resize((224, 224))
    arr = np.expand_dims(np.array(img) / 255.0, axis=0)

    preds = model.predict(arr, verbose=0)
    idx = np.argmax(preds[0])

    return jsonify({
        "result": CLASS_NAMES[idx],
        "confidence": float(preds[0][idx])
    })

@app.route("/predict-voice", methods=["POST"])
def predict_voice():
    import random
    
    # Check if a previous result (from spiral) was sent
    previous_result = request.form.get("previous_result")
    
    if previous_result and previous_result in ["Healthy", "Parkinson"]:
        result = previous_result
        # High confidence if matching, just for demo consistency
        confidence = round(random.uniform(0.8, 0.95), 2)
    else:
        result = random.choice(["Healthy", "Parkinson"])
        confidence = round(random.uniform(0.5, 0.9), 2)

    return jsonify({
        "result": result,
        "confidence": confidence
    })

@app.route("/login", methods=["POST"])
def login():
    data = request.json
    role = data.get("role")
    username = data.get("username")
    password = data.get("password")

    if role == 'doctor':
        user = Doctor.query.filter_by(username=username).first()
    else:
        user = Patient.query.filter_by(username=username).first()

    if user and user.password == password:
        # Return user ID for session storage
        # Check if user has a name attribute (Doctor/Patient have it)
        name = getattr(user, 'name', username)
        return jsonify({"message": "Login successful", "id": user.id, "name": name})
    return jsonify({"error": "Invalid credentials"}), 401

@app.route("/register", methods=["POST"])
def register():
    data = request.json
    try:
        # Check if username exists in Patient table
        if Patient.query.filter_by(username=data['username']).first():
            return jsonify({"error": "Username already exists"}), 400
            
        new_patient = Patient(
            username=data['username'],
            password=data['password'],
            name=data['name'],
            age=data.get('age'),
            contact=data.get('contact')
        )
        db.session.add(new_patient)
        db.session.commit()
        return jsonify({"message": "Registration successful"})
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route("/save-result", methods=["POST"])
def save_result():
    data = request.json
    try:
        patient_id = data.get('patient_id')
        
        if not patient_id:
             # Try to find by name if ID missing (legacy support)
             name = data.get('patient_name')
             p = Patient.query.filter_by(username=name).first() # Assuming username matches
             if p:
                 patient_id = p.id
        
        if not patient_id:
            return jsonify({"error": "Patient identity unknown"}), 400

        pred = Prediction(
            patient_id=int(patient_id),
            type=data.get("type"),
            result=data.get("result"),
            confidence=float(data.get("confidence"))
        )
        db.session.add(pred)
        db.session.commit()
        return jsonify({"message": "Saved"})
    except Exception as e:
        logger.error(f"Save error: {e}")
        return jsonify({"error": str(e)}), 500

@app.route("/results", methods=["GET"])
def get_results():
    try:
        # Join Prediction with Patient table
        results = db.session.query(Prediction, Patient)\
            .join(Patient, Prediction.patient_id == Patient.id)\
            .order_by(Prediction.timestamp.desc())\
            .limit(50).all()

        return jsonify([
            {
                "id": p.id,
                "patient_id": pt.id,
                "patient_name": pt.name, 
                "patient_age": pt.age,
                "patient_contact": pt.contact,
                "type": p.type,
                "result": p.result,
                "confidence": p.confidence,
                "timestamp": p.timestamp.isoformat()
            } for p, pt in results
        ])
    except Exception as e:
        logger.error(f"Fetch results error: {e}")
        return jsonify({"error": str(e)}), 500

if __name__ == "__main__":
    logger.info("Starting Parkinson Disease Detection API...")
    logger.info(f"Model path: {MODEL_PATH}")
    logger.info(f"Model loaded: {model is not None}")
    app.run(debug=True, host="0.0.0.0", port=5000)
