backend/requirements.txt
flask
tensorflow
numpy
pillow
flask-cors
librosa
soundfile
flask-sqlalchemy
mysql-connector-python
